chore: move themeConfig to config directory
